refactor: move Page component into its own file
