+++ .changeset/tall-crabs-bake.md
@@ -1,0 +1,5 @@
+---
+'@makeswift/runtime': patch
+---
+
+Move `Page` component into its own file.
